Yogendra

+++ DataTransferObject/Domain/Master/MAppointment.cs
@@ -12,7 +12,7 @@ namespace DataTransferObject.Domain.Master
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public byte ApptId { get; set; }
+        public short ApptId { get; set; }
         [Column(TypeName = "VARCHAR(50)")]
         [Required(ErrorMessage = "required!")]
         [RegularExpression("^[a-zA-Z0-9 ]*$", ErrorMessage = "Only Alphabets and Numbers allowed.")]
